+++ framework/docs/src/releases/CHANGELOG.md
@@ -6,6 +6,8 @@
 
 - `execute_with_funds` to Executor to attach funds to execution.
 - `stargate` feature for abstract-app, abstract-standalone and abstract-adapter packages.
+- New module type: `Service`, behaves the same as Native, but can be registered by any namespace.
+- `AbstractClient`: `service` to get api of Service module
 
 ### Changed
 
@@ -22,8 +24,11 @@
 - `enable_ibc` added to Account builder.
 - `module_status` on AbstractClient that returns current status of the module.
 - `install_on_sub_account` now defaults to `false` in Account Builder
+- `Publisher` will check if dependencies of the module is registered in version control to the chain before publishing.
 
 ### Removed
+
+- `cw-semver` dependency removed
 
 ### Fixed